Separate `WebAwesomeFormAssociatedElement` (and friends) into a separate file
