+++ src/internal/test/form-control-base-tests.ts
@@ -2,7 +2,7 @@ import { aTimeout, expect } from '@open-wc/testing';
 import { html, type TemplateResult } from 'lit';
 import { html as staticHTML, unsafeStatic } from 'lit/static-html.js';
 import { clickOnElement } from '../test.js';
-import type { WebAwesomeFormControl } from '../webawesome-element.js';
+import type { WebAwesomeFormControl } from '../webawesome-formassociated-element.js';
 import type { clientFixture, hydratedFixture } from './fixture.js';
 import { fixtures } from './fixture.js';
